+++ s4e-backend/src/main/java/pl/cyfronet/s4e/Constants.java
@@ -2,7 +2,6 @@ package pl.cyfronet.s4e;
 
 public final class Constants {
     public static final String API_PREFIX_V1 = "/api/v1";
-    public static final String API_PREFIX_S3 = "/api/s3";
     public static final String GEOSERVER_PRG_DATA_STORE = "prg";
     public static final String GEOSERVER_PRG_PATH = "/opt/geoserver/prg/";
 }