+++ s4e-backend/src/main/java/pl/cyfronet/s4e/config/PersistenceConfig.java
@@ -4,12 +4,9 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContext;
-import org.springframework.security.core.context.SecurityContextHolder;
+import pl.cyfronet.s4e.security.AppUserDetails;
+import pl.cyfronet.s4e.util.AppUserDetailsSupplier;
 
-import java.security.Principal;
-import java.util.Objects;
 import java.util.Optional;
 
 @Configuration
@@ -17,10 +14,7 @@ import java.util.Optional;
 public class PersistenceConfig {
     @Bean
     public AuditorAware<String> auditorAware() {
-        return () -> Optional.ofNullable(SecurityContextHolder.getContext())
-                .map(SecurityContext::getAuthentication)
-                .filter(Objects::nonNull)
-                .filter(Authentication::isAuthenticated)
-                .map(Principal::getName);
+        return () -> Optional.ofNullable(AppUserDetailsSupplier.get())
+                .map(AppUserDetails::getUsername);
     }
 }