+++ s4e-backend/src/main/java/pl/cyfronet/s4e/config/ThreadConfig.java
@@ -14,7 +14,6 @@ import java.util.concurrent.Executor;
 @Configuration
 @EnableAsync
 public class ThreadConfig implements AsyncConfigurer {
-
     @Override
     public Executor getAsyncExecutor() {
         ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();